[REFACTOR] PropertyConvertContext 도입
역시나 컨버터별로 요구되는 함수 인자가 달랐기에 도입하였음
