+++ src/main/java/org/hanihome/hanihomebe/property/application/converter/PropertyDefaultConverter.java
@@ -1,4 +1,5 @@
-package org.hanihome.hanihomebe.property.application;
+package org.hanihome.hanihomebe.property.application.converter;
+
 
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
@@ -8,37 +9,40 @@ import org.hanihome.hanihomebe.property.domain.RentProperty;
 import org.hanihome.hanihomebe.property.domain.ShareProperty;
 import org.hanihome.hanihomebe.property.domain.enums.PropertySuperType;
 import org.hanihome.hanihomebe.property.web.dto.enums.PropertyViewType;
-import org.hanihome.hanihomebe.property.web.dto.response.summary.PropertySummaryDTO;
-import org.hanihome.hanihomebe.property.web.dto.response.summary.RentPropertySummaryDTO;
-import org.hanihome.hanihomebe.property.web.dto.response.summary.SharePropertySummaryDTO;
+import org.hanihome.hanihomebe.property.web.dto.response.PropertyResponseDTO;
+import org.hanihome.hanihomebe.property.web.dto.response.RentPropertyResponseDTO;
+import org.hanihome.hanihomebe.property.web.dto.response.SharePropertyResponseDTO;
 import org.hibernate.Hibernate;
 import org.hibernate.proxy.HibernateProxy;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+
 @Component
-public class PropertySummaryConverter implements PropertyConverter<PropertySummaryDTO>{
+public class PropertyDefaultConverter implements PropertyConverter<PropertyResponseDTO> {
     @Override
     public PropertyViewType supports() {
-        return PropertyViewType.SUMMARY;
+        return PropertyViewType.DEFAULT;
     }
 
     @Override
-    public PropertySummaryDTO convert(Property property, List<OptionItemResponseDTO> optionItems) {
-        return toSummaryDTO(property);
+    public PropertyResponseDTO convert(PropertyConvertContext propertyConvertContext) {
+        return toResponseDTO(propertyConvertContext.getProperty(), propertyConvertContext.getOptionItems());
     }
 
-    PropertySummaryDTO toSummaryDTO(Property entity) {
+
+    PropertyResponseDTO toResponseDTO(Property entity, List<OptionItemResponseDTO> optionItems) {
         PropertySuperType propertyType = entity.getKind();
         switch (propertyType) {
             case SHARE:
-                return SharePropertySummaryDTO.from(safeCast(entity, ShareProperty.class));
+                return SharePropertyResponseDTO.from(safeCast(entity, ShareProperty.class), optionItems);
             case RENT:
-                return RentPropertySummaryDTO.from(safeCast(entity, RentProperty.class));
+                return RentPropertyResponseDTO.from(safeCast(entity, RentProperty.class), optionItems);
             default:
                 throw new CustomException(ServiceCode.INVALID_PROPERTY_TYPE);
         }
     }
+
     private <T> T safeCast(Object entity, Class<T> targetClass) {
         if (entity instanceof HibernateProxy) {
             return targetClass.cast(Hibernate.unproxy(entity));
