+++ src/main/java/org/hanihome/hanihomebe/property/application/converter/PropertySummaryConverter.java
@@ -1,18 +1,13 @@
-package org.hanihome.hanihomebe.property.application;
-
+package org.hanihome.hanihomebe.property.application.converter;
 
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
-import org.hanihome.hanihomebe.item.web.dto.OptionItemResponseDTO;
-import org.hanihome.hanihomebe.property.application.PropertyConverter;
+import org.hanihome.hanihomebe.metro.web.dto.nearest.NearestMetroStopResponseDTO;
 import org.hanihome.hanihomebe.property.domain.Property;
 import org.hanihome.hanihomebe.property.domain.RentProperty;
 import org.hanihome.hanihomebe.property.domain.ShareProperty;
 import org.hanihome.hanihomebe.property.domain.enums.PropertySuperType;
 import org.hanihome.hanihomebe.property.web.dto.enums.PropertyViewType;
-import org.hanihome.hanihomebe.property.web.dto.response.PropertyResponseDTO;
-import org.hanihome.hanihomebe.property.web.dto.response.RentPropertyResponseDTO;
-import org.hanihome.hanihomebe.property.web.dto.response.SharePropertyResponseDTO;
 import org.hanihome.hanihomebe.property.web.dto.response.summary.PropertySummaryDTO;
 import org.hanihome.hanihomebe.property.web.dto.response.summary.RentPropertySummaryDTO;
 import org.hanihome.hanihomebe.property.web.dto.response.summary.SharePropertySummaryDTO;
@@ -20,33 +15,30 @@ import org.hibernate.Hibernate;
 import org.hibernate.proxy.HibernateProxy;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
-
 @Component
-public class PropertyDefaultConverter implements PropertyConverter<PropertyResponseDTO> {
+public class PropertySummaryConverter implements PropertyConverter<PropertySummaryDTO>{
     @Override
     public PropertyViewType supports() {
-        return PropertyViewType.DEFAULT;
+        return PropertyViewType.SUMMARY;
     }
 
     @Override
-    public PropertyResponseDTO convert(Property property, List<OptionItemResponseDTO> optionItems) {
-        return toResponseDTO(property, optionItems);
+    public PropertySummaryDTO convert(PropertyConvertContext propertyConvertContext) {
+        return toSummaryDTO(propertyConvertContext.getProperty(), propertyConvertContext.getNearestMetroStopResponseDTO());
     }
 
-
-    PropertyResponseDTO toResponseDTO(Property entity, List<OptionItemResponseDTO> optionItems) {
+    PropertySummaryDTO toSummaryDTO(Property entity, NearestMetroStopResponseDTO nearestMetroStopResponseDTO) {
         PropertySuperType propertyType = entity.getKind();
         switch (propertyType) {
             case SHARE:
-                return SharePropertyResponseDTO.from(safeCast(entity, ShareProperty.class), optionItems);
+                return SharePropertySummaryDTO.from(safeCast(entity, ShareProperty.class), nearestMetroStopResponseDTO);
             case RENT:
-                return RentPropertyResponseDTO.from(safeCast(entity, RentProperty.class), optionItems);
+                return RentPropertySummaryDTO.from(safeCast(entity, RentProperty.class), nearestMetroStopResponseDTO);
             default:
                 throw new CustomException(ServiceCode.INVALID_PROPERTY_TYPE);
         }
-    }
 
+    }
     private <T> T safeCast(Object entity, Class<T> targetClass) {
         if (entity instanceof HibernateProxy) {
             return targetClass.cast(Hibernate.unproxy(entity));
